requirements.txt
beautifulsoup4
lxml
wheel
requests
coverage
requests[socks]
flake8
celery
redis
pyyaml